requirements.txt
fastapi>=0.103.1
openai-agents==0.0.13
openai==1.76.0
pre-commit==3.7.1
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart>=0.0.6
runwayml==3.0.3
uvicorn>=0.23.2
matplotlib==3.10.1
Pillow>=9.0.0
